Revert "[Refactor] separate `RecyclableClient` class from `HiveMetaClient` (#29861)"
This reverts commit e73e202d5ccd207e5554375d3db3eb5ac2dd5233.
